refactor: 외부 라이브러리 사용 객체의 인터페이스 추가 및 관련 객체 수정
